Refactor experimental modules according to the SmallRye definitions
