+++ testsuite/experimental/src/test/java/io/smallrye/health/SmallRyeBeanArchiveHandler.java
