+++ testsuite/experimental/src/test/java/io/smallrye/health/SmallRyeHealthArchiveProcessor.java
