+++ testsuite/experimental/src/test/java/io/smallrye/health/deployment/SuccessfulCustom.java
@@ -26,7 +26,7 @@ import javax.enterprise.context.ApplicationScoped;
 import org.eclipse.microprofile.health.HealthCheck;
 import org.eclipse.microprofile.health.HealthCheckResponse;
 
-import io.smallrye.health.HealthGroup;
+import io.smallrye.health.api.HealthGroup;
 
 @HealthGroup("group1")
 @HealthGroup("group2")
