+++ testsuite/experimental/src/test/resources/META-INF/beans.xml
