+++ testsuite/experimental/src/test/resources/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
