+++ testsuite/pom.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <parent>
+    <artifactId>smallrye-health-parent</artifactId>
+    <groupId>io.smallrye</groupId>
+    <version>2.2.2-SNAPSHOT</version>
+  </parent>
+  <modelVersion>4.0.0</modelVersion>
+
+  <artifactId>smallrye-health-testsuite-parent</artifactId>
+  <name>SmallRye Health: Test Suite Parent</name>
+  <packaging>pom</packaging>
+  
+  <modules>
+    <module>experimental</module>
+    <module>tck</module>
+  </modules>
+
+
+</project>
