+++ testsuite/tck/src/test/java/io/smallrye/health/SmallRyeBeanArchiveHandler.java
