+++ testsuite/tck/src/test/resources/META-INF/beans.xml
