+++ testsuite/tck/src/test/resources/arquillian.xml
